# openapi.yaml
openapi: 3.0.0
info:
  title: Learnify API
  version: 1.0.0
  description: |
    Learnify is a personalized learning platform API that provides endpoints for user authentication,
    course management, exercise submissions, social features, and AI-powered code reviews.

    ## Authentication
    All protected endpoints require a JWT Bearer token in the Authorization header:
    `Authorization: Bearer <token>`

    ## Features
    - User registration and authentication
    - Personalized learning paths
    - AI-powered code reviews
    - Social features (follow, activity feed)
    - Achievement system
    - Course recommendations

  contact:
    name: Learnify API Support
    email: support@learnify.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.learnify.example.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request body"
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        avatar_url:
          type: string
          format: uri
          example: "https://example.com/avatars/user123.jpg"
        privacy_settings:
          $ref: '#/components/schemas/PrivacySettings'
        created_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-20T14:45:00Z"
        last_login:
          type: string
          format: date-time
          example: "2025-01-21T09:00:00Z"
    PrivacySettings:
      type: object
      properties:
        profile_visibility:
          type: string
          enum: [public, friends, private]
          example: "public"
        activity_visibility:
          type: string
          enum: [public, friends, private]
          example: "friends"
        progress_visibility:
          type: string
          enum: [public, friends, private]
          example: "public"
        allow_followers:
          type: boolean
          example: true
        show_in_leaderboards:
          type: boolean
          example: true
        show_completed_courses:
          type: boolean
          example: true
...
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          example: "John Updated"
        avatar_url:
          type: string
          format: uri
          example: "https://example.com/avatars/new-avatar.jpg"
...
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "456e7890-e89b-12d3-a456-426614174001"
        user_id:
          type: string
          format: uuid
        archetype_id:
          type: string
          format: uuid
        title:
          type: string
          example: "Modern Web Development with React"
        description:
          type: string
          example: "Learn to build modern web applications using React, Node.js, and PostgreSQL"
        meta_category:
          type: string
          example: "programming"
        status:
          type: string
          enum: [active, completed, archived]
          example: "active"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ActivityFeed:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        activity_type:
          type: string
          enum: [course_started, module_completed, exercise_submitted, achievement_unlocked, user_followed]
          example: "module_completed"
        reference_type:
          type: string
          example: "module"
        reference_id:
          type: string
          format: uuid
        metadata:
          type: object
          example:
            module_title: "Introduction to React Components"
            course_title: "Modern Web Development"
        visibility:
          type: string
          enum: [public, friends, private]
          example: "public"
        created_at:
          type: string
          format: date-time
    Achievement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "First Steps"
        description:
          type: string
          example: "Complete your first exercise"
        badge_icon:
          type: string
          format: uri
          example: "https://example.com/badges/first-steps.png"
        rarity:
          type: string
          enum: [common, rare, epic, legendary]
          example: "common"
        unlocked_at:
          type: string
          format: date-time
          description: When the user unlocked this achievement
    Recommendation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        recommendation_type:
          type: string
          enum: [collaborative_filtering, skill_adjacency, social_signal, trending]
          example: "skill_adjacency"
        match_score:
          type: integer
          minimum: 0
          maximum: 100
          example: 87
        reason:
          type: string
          example: "Based on your JavaScript skills, you'll love TypeScript"
        course:
          $ref: '#/components/schemas/Course'
    TrendingCourse:
      type: object
      properties:
        course_id:
          type: string
          format: uuid
        title:
          type: string
          example: "Advanced React Patterns"
        velocity:
          type: number
          format: float
          example: 2.5
          description: Growth rate compared to previous period
        signups_24h:
          type: integer
          example: 156
        rank:
          type: integer
          example: 1
        meta_category:
          type: string
          example: "programming"
paths:
...
  /api/users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Retrieves the authenticated user's profile information
      operationId: getUserProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Users
      summary: Update user profile
      description: Updates the authenticated user's profile (name, avatar)
      operationId: updateUserProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "profile updated successfully"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/feed:
    get:
      tags:
        - Social
      summary: Get activity feed
      description: Retrieves personalized activity feed based on followed users
      operationId: getActivityFeed
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of activities to return
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Activity feed retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  activities:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityFeed'
                  count:
                    type: integer
                    example: 25
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example: "Unauthorized"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
  /api/users/{id}/follow:
    post:
      tags:
        - Social
      summary: Follow a user
      description: Creates a follow relationship with another user
      operationId: followUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User UUID to follow
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Successfully followed user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully followed user"
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                type: string
                example: "User ID is required"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example: "Unauthorized"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: string

    delete:
      tags:
        - Social
      summary: Unfollow a user
      description: Removes a follow relationship with another user
      operationId: unfollowUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User UUID to unfollow
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully unfollowed user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successfully unfollowed user"
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                type: string
                example: "User ID is required"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example: "Unauthorized"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
  /api/recommendations:
    get:
      tags:
        - Social
      summary: Get course recommendations
      description: Retrieves personalized course recommendations grouped by type
      operationId: getRecommendations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Recommendations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recommendation'
                  sections:
                    type: object
                    properties:
                      collaborative_filtering:
                        type: string
                        example: "Because You Completed"
                      skill_adjacency:
                        type: string
                        example: "Next Level Skills"
                      social_signal:
                        type: string
                        example: "Friends Are Learning"
                      trending:
                        type: string
                        example: "Trending Now"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example: "Unauthorized"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
  /api/trending:
    get:
      tags:
        - Social
      summary: Get trending courses
      description: Retrieves currently trending courses (public endpoint)
      operationId: getTrendingCourses
      responses:
        '200':
          description: Trending courses retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  trending:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrendingCourse'
                  count:
                    type: integer
                    example: 10
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
  /api/users/{id}/profile:
    get:
      tags:
        - Social
      summary: Get user profile (Living Resume)
      description: Retrieves comprehensive user profile with achievements and progress
      operationId: getUserPublicProfile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  stats:
                    type: object
                    properties:
                      total_courses:
                        type: integer
                      completed_courses:
                        type: integer
                      total_exercises:
                        type: integer
                      achievements_count:
                        type: integer
                  recent_activity:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityFeed'
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
  /api/users/me/achievements:
    get:
      tags:
        - Achievements
      summary: Get user achievements
      description: Retrieves all achievements unlocked by the authenticated user
      operationId: getUserAchievements
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Achievements retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  achievements:
                    type: array
                    items:
                      $ref: '#/components/schemas/Achievement'
                  count:
                    type: integer
                    example: 5
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example: "Unauthorized"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: string
...